Brugervenlige adminfiltre (#387)
* Gør listefiltre mere brugervenlige ved sortering alfabetisk efter afdeling, evt efterfulgt at aktivitetsstartdato
