+++ members/admin/__init__.py
@@ -2,41 +2,39 @@ from uuid import uuid4
 from django import forms
 from django.contrib import admin
 from django.db import models
-from django.db import transaction
 from django.db.models import Q
-from members.models.person import Person
-from members.models.department import Department
-from members.models.union import Union
-from members.models.volunteer import Volunteer
-from members.models.member import Member
-from members.models.activity import Activity
-from members.models.activityinvite import ActivityInvite
-from members.models.activityparticipant import ActivityParticipant
-from members.models.family import Family
-from members.models.emailitem import EmailItem
-from members.models.waitinglist import WaitingList
-from members.models.emailtemplate import EmailTemplate
-from members.models.payment import Payment
-from members.models.equipment import Equipment
-from members.models.equipmentloan import EquipmentLoan
+
 from django.contrib.auth.models import User
 from django.db.models.functions import Lower
-from django.http import HttpResponse
 from django.utils import timezone
-from datetime import timedelta
 from django.urls import reverse
 from django.utils.html import format_html
 from django.forms import Textarea
-from django.shortcuts import render
-from django.contrib.admin.widgets import AdminDateWidget
-from django.contrib import messages
 
-from members.models import Address, AdminUserInformation
+from members.models import (
+    Address,
+    AdminUserInformation,
+    Person,
+    Department,
+    Union,
+    Volunteer,
+    Member,
+    Activity,
+    ActivityInvite,
+    ActivityParticipant,
+    Family,
+    EmailItem,
+    Payment,
+    Equipment,
+    EquipmentLoan,
+)
 
 from .address_admin import AddressAdmin
 from .department_admin import DepartmentAdmin
 from .union_admin import UnionAdmin
 from .user_admin import UserAdmin
+from .person_admin import PersonAdmin
+from .member_admin import MemberAdmin
 
 admin.site.site_header = "Coding Pirates Medlemsdatabase"
 admin.site.index_title = "Afdelings admin"
@@ -46,6 +44,8 @@ admin.site.register(Department, DepartmentAdmin)
 admin.site.register(Union, UnionAdmin)
 admin.site.unregister(User)
 admin.site.register(User, UserAdmin)
+admin.site.register(Person, PersonAdmin)
+admin.site.register(Member, MemberAdmin)
 
 
 class EmailItemInline(admin.TabularInline):
@@ -61,28 +61,6 @@ class EmailItemInline(admin.TabularInline):
         return False
 
     extra = 0
-
-
-class MemberAdmin(admin.ModelAdmin):
-    list_display = ("name", "department", "member_since", "is_active")
-    list_filter = ["department"]
-    list_per_page = 20
-    raw_id_fields = ("department", "person")
-
-    # Only view mebers related to users department
-    def get_queryset(self, request):
-        qs = super(MemberAdmin, self).get_queryset(request)
-        if request.user.is_superuser:
-            return qs
-        departments = Department.objects.filter(
-            adminuserinformation__user=request.user
-        ).values("id")
-        return qs.filter(
-            activityparticipant__activity__department__in=departments
-        ).distinct()
-
-
-admin.site.register(Member, MemberAdmin)
 
 
 class ActivityAdmin(admin.ModelAdmin):
@@ -326,11 +304,11 @@ class ParticipantPaymentListFilter(admin.SimpleListFilter):
         """
 
         activitys = [
-            ("none", "Ikke betalt"),
-            ("ok", "Betalt"),
-            ("confirmed", "Hævet"),
             ("pending", "Afventende"),
             ("rejected", "Afvist"),
+            ("ok", "Betalt"),
+            ("none", "Ikke betalt"),
+            ("confirmed", "Hævet"),
         ]
         return activitys
 
@@ -374,7 +352,7 @@ class ActivityParticipantListFilter(admin.SimpleListFilter):
         activitys = []
         for activity in Activity.objects.filter(
             department__in=AdminUserInformation.get_departments_admin(request.user)
-        ):
+        ).order_by("department__name", "-start_date"):
             activitys.append((str(activity.pk), str(activity)))
         return activitys
 
@@ -434,7 +412,7 @@ class ActivivtyInviteActivityListFilter(admin.SimpleListFilter):
         activitys = []
         for activity in Activity.objects.filter(
             department__in=AdminUserInformation.get_departments_admin(request.user)
-        ).order_by("zipcode"):
+        ).order_by("department__name"):
             activitys.append((str(activity.pk), activity))
 
         return activitys
@@ -508,514 +486,6 @@ class ActivityInviteAdmin(admin.ModelAdmin):
 
 
 admin.site.register(ActivityInvite, ActivityInviteAdmin)
-
-
-class PersonWaitinglistListFilter(admin.SimpleListFilter):
-    title = "Venteliste"
-    parameter_name = "waiting_list"
-
-    def lookups(self, request, model_admin):
-
-        departments = [
-            ("any", "Alle opskrevne samlet"),
-            ("none", "Ikke skrevet på venteliste"),
-        ]
-        for department in AdminUserInformation.get_departments_admin(request.user):
-            departments.append((str(department.pk), department.name))
-
-        return departments
-
-    def queryset(self, request, queryset):
-        """
-        Returns the filtered queryset based on the value
-        provided in the query string and retrievable via
-        `self.value()`.
-        """
-        # Compare the requested value (either '80s' or '90s')
-        # to decide how to filter the queryset.
-
-        if self.value() == "any":
-            return queryset.exclude(waitinglist__isnull=True)
-        elif self.value() == "none":
-            return queryset.filter(waitinglist__isnull=True)
-        elif self.value() is None:
-            return queryset
-        else:
-            return queryset.filter(waitinglist__department__pk=self.value())
-
-
-class VolunteerListFilter(admin.SimpleListFilter):
-    title = "frivillig i"
-    parameter_name = "volunteer"
-
-    def lookups(self, request, model_admin):
-        departments = [("any", "Alle frivillige samlet"), ("none", "Ikke frivillig")]
-        for department in AdminUserInformation.get_departments_admin(request.user):
-            departments.append((str(department.pk), department.name))
-
-        return departments
-
-    def queryset(self, request, queryset):
-        """
-        Returns the filtered queryset based on the value
-        provided in the query string and retrievable via
-        `self.value()`.
-        """
-        # Compare the requested value (either '80s' or '90s')
-        # to decide how to filter the queryset.
-
-        if self.value() == "any":
-            return (
-                queryset.filter(volunteer__isnull=False)
-                .filter(volunteer__removed__isnull=True)
-                .distinct()
-            )
-        elif self.value() == "none":
-            return (
-                queryset.filter(volunteer__isnull=True).distinct()
-                | queryset.exclude(volunteer__removed__isnull=True).distinct()
-            )
-        elif self.value() is None:
-            return queryset
-        else:
-            return queryset.filter(
-                volunteer__department__pk=self.value(), volunteer__removed__isnull=True
-            )
-
-
-class PersonParticipantListFilter(admin.SimpleListFilter):
-    title = "Deltager på"
-    parameter_name = "participant_list"
-
-    def lookups(self, request, model_admin):
-
-        activitys = [("none", "Deltager ikke"), ("any", "Alle deltagere samlet")]
-        for activity in (
-            Activity.objects.filter(
-                department__in=AdminUserInformation.get_departments_admin(request.user)
-            )
-            .order_by("start_date")
-            .order_by("zipcode")
-        ):
-            activitys.append((str(activity.pk), str(activity)))
-
-        return activitys
-
-    def queryset(self, request, queryset):
-        if self.value() == "none":
-            return queryset.filter(member__activityparticipant__isnull=True)
-        elif self.value() == "any":
-            return queryset.exclude(member__activityparticipant__isnull=True)
-        elif self.value() is None:
-            return queryset
-        else:
-            return queryset.filter(member__activityparticipant__activity=self.value())
-
-
-class PersonInvitedListFilter(admin.SimpleListFilter):
-    title = "Inviteret til"
-    parameter_name = "activity_invited_list"
-
-    def lookups(self, request, model_admin):
-        activitys = [("none", "Ikke inviteret til noget"), ("any", "Alle inviterede")]
-        for activity in (
-            Activity.objects.filter(
-                department__in=AdminUserInformation.get_departments_admin(request.user)
-            )
-            .order_by("start_date")
-            .order_by("zipcode")
-        ):
-            activitys.append((str(activity.pk), str(activity)))
-
-        return activitys
-
-    def queryset(self, request, queryset):
-        """
-        Returns the filtered queryset based on the value
-        provided in the query string and retrievable via
-        `self.value()`.
-        """
-        # Compare the requested value (either '80s' or '90s')
-        # to decide how to filter the queryset.
-
-        if self.value() == "none":
-            return queryset.filter(activityinvite__isnull=True)
-        elif self.value() == "any":
-            return queryset.exclude(activityinvite__isnull=True)
-        elif self.value() is None:
-            return queryset
-        else:
-            return queryset.filter(activityinvite__activity=self.value())
-
-
-class WaitingListInline(admin.TabularInline):
-    model = WaitingList
-    fields = ["on_waiting_list_since", "department", "number_on_waiting_list"]
-    readonly_fields = fields
-    extra = 0
-
-
-class PersonAdmin(admin.ModelAdmin):
-    list_display = (
-        "name",
-        "membertype",
-        "family_url",
-        "age_years",
-        "zipcode",
-        "added",
-        "notes",
-    )
-    list_filter = (
-        "membertype",
-        "gender",
-        VolunteerListFilter,
-        PersonWaitinglistListFilter,
-        PersonInvitedListFilter,
-        PersonParticipantListFilter,
-    )
-    search_fields = ("name", "family__email", "notes")
-    actions = ["invite_many_to_activity_action", "export_emaillist", "export_csv"]
-    raw_id_fields = ("family", "user")
-
-    inlines = [
-        PaymentInline,
-        VolunteerInline,
-        ActivityInviteInline,
-        MemberInline,
-        WaitingListInline,
-    ]
-
-    def family_url(self, item):
-        return format_html(
-            '<a href="../family/%d">%s</a>' % (item.family.id, item.family.email)
-        )
-
-    family_url.allow_tags = True
-    family_url.short_description = "Familie"
-    list_per_page = 20
-
-    def invite_many_to_activity_action(self, request, queryset):
-        # Get list of available departments
-        if request.user.is_superuser or request.user.has_perm(
-            "members.view_all_persons"
-        ):
-            deparment_list_query = Department.objects.all()
-        else:
-            deparment_list_query = Department.objects.filter(
-                adminuserinformation__user=request.user
-            )
-        deparment_list = [("-", "-")]
-        for department in deparment_list_query:
-            deparment_list.append((department.id, department.name))
-
-        # Get list of active and future activities
-        department_ids = deparment_list_query.values_list("id", flat=True)
-        activity_list_query = Activity.objects.filter(end_date__gt=timezone.now())
-        if not request.user.is_superuser:
-            activity_list_query = activity_list_query.filter(
-                department__in=department_ids
-            )
-        activity_list = [("-", "-")]
-        for activity in activity_list_query:
-            activity_list.append(
-                (activity.id, activity.department.name + ", " + activity.name)
-            )
-
-        # Form used to select department and activity - redundant department is for double check
-        class MassInvitationForm(forms.Form):
-            department = forms.ChoiceField(label="Afdeling", choices=deparment_list)
-            activity = forms.ChoiceField(label="Aktivitet", choices=activity_list)
-            expire = forms.DateField(
-                label="Udløber",
-                widget=AdminDateWidget(),
-                initial=timezone.now() + timedelta(days=30 * 3),
-            )
-
-        # Lookup all the selected persons - to show confirmation list
-        persons = queryset
-
-        context = admin.site.each_context(request)
-        context["persons"] = persons
-        context["queryset"] = queryset
-
-        if request.method == "POST" and "department" in request.POST:
-            # Post request with data
-            mass_invitation_form = MassInvitationForm(request.POST)
-            context["mass_invitation_form"] = mass_invitation_form
-
-            if (
-                mass_invitation_form.is_valid()
-                and mass_invitation_form.cleaned_data["activity"] != "-"
-                and mass_invitation_form.cleaned_data["department"] != "-"
-            ):
-                activity = Activity.objects.get(
-                    pk=mass_invitation_form.cleaned_data["activity"]
-                )
-
-                # validate activity belongs to user and matches selected department
-                if (
-                    int(mass_invitation_form.cleaned_data["department"])
-                    in department_ids
-                ):
-                    if activity.department.id == int(
-                        mass_invitation_form.cleaned_data["department"]
-                    ):
-                        invited_counter = 0
-
-                        # get list of already created invitations on selected persons
-                        already_invited = Person.objects.filter(
-                            activityinvite__activity=mass_invitation_form.cleaned_data[
-                                "activity"
-                            ],
-                            activityinvite__person__in=queryset,
-                        ).all()
-                        list(already_invited)  # force lookup
-                        already_invited_ids = already_invited.values_list(
-                            "id", flat=True
-                        )
-
-                        # only save if all succeeds
-                        try:
-                            with transaction.atomic():
-                                for current_person in queryset:
-                                    if (
-                                        current_person.id not in already_invited_ids
-                                        and (
-                                            activity.max_age
-                                            >= current_person.age_years()
-                                            >= activity.min_age
-                                        )
-                                    ):
-                                        invited_counter = invited_counter + 1
-                                        invitation = ActivityInvite(
-                                            activity=activity,
-                                            person=current_person,
-                                            expire_dtm=mass_invitation_form.cleaned_data[
-                                                "expire"
-                                            ],
-                                        )
-                                        invitation.save()
-                        except Exception:
-                            messages.error(
-                                request,
-                                "Fejl - ingen personer blev inviteret! Der var problemer med "
-                                + invitation.person.name
-                                + ". Vær sikker på personen ikke allerede er inviteret og opfylder alderskravet.",
-                            )
-                            return
-
-                        # return ok message
-                        already_invited_text = ""
-                        if already_invited.count():
-                            already_invited_text = (
-                                ". Dog var : "
-                                + str.join(
-                                    ", ", already_invited.values_list("name", flat=True)
-                                )
-                                + " allerede inviteret!"
-                            )
-                        messages.success(
-                            request,
-                            str(invited_counter)
-                            + " af "
-                            + str(queryset.count())
-                            + " valgte personer blev inviteret til "
-                            + str(activity)
-                            + already_invited_text,
-                        )
-                        return
-
-                    else:
-                        messages.error(
-                            request,
-                            "Valgt aktivitet stemmer ikke overens med valgt afdeling",
-                        )
-                        return
-                else:
-                    messages.error(request, "Du kan kun invitere til egne afdelinger")
-                    return
-        else:
-            context["mass_invitation_form"] = MassInvitationForm()
-
-        return render(request, "admin/invite_many_to_activity.html", context)
-
-    invite_many_to_activity_action.short_description = (
-        "Inviter alle valgte til en aktivitet"
-    )
-
-    # needs 'view_full_address' to set personal details.
-    # email and phonenumber only shown on adults.
-    def get_fieldsets(self, request, person=None):
-        if request.user.has_perm("members.view_full_address"):
-            contact_fields = (
-                "name",
-                "streetname",
-                "housenumber",
-                "floor",
-                "door",
-                "city",
-                "zipcode",
-                "placename",
-                "email",
-                "phone",
-                "family",
-            )
-        else:
-            if person.membertype == Person.CHILD:
-                contact_fields = ("name", "city", "zipcode", "family")
-            else:
-                contact_fields = ("name", "city", "zipcode", "email", "phone", "family")
-
-        fieldsets = (
-            ("Kontakt Oplysninger", {"fields": contact_fields}),
-            ("Noter", {"fields": ("notes",)}),
-            (
-                "Yderlige informationer",
-                {
-                    "classes": ("collapse",),
-                    "fields": (
-                        "membertype",
-                        "birthday",
-                        "has_certificate",
-                        "added",
-                        "user",
-                    ),
-                },
-            ),
-        )
-
-        return fieldsets
-
-    def get_readonly_fields(self, request, obj=None):
-        if type(obj) == Person and not request.user.is_superuser:
-            return [
-                "name",
-                "streetname",
-                "housenumber",
-                "floor",
-                "door",
-                "city",
-                "zipcode",
-                "placename",
-                "email",
-                "phone",
-                "family",
-                "membertype",
-                "birthday",
-                "has_certificate",
-                "added",
-            ]
-        else:
-            return []
-
-    def unique(self, item):
-        return item.family.unique if item.family is not None else ""
-
-    def export_emaillist(self, request, queryset):
-        result_string = "kopier denne liste direkte ind i dit email program (Husk at bruge Bcc!)\n\n"
-        family_email = []
-        for person in queryset:
-            if not person.family.dont_send_mails:
-                family_email.append(person.family.email)
-        result_string = result_string + ";\n".join(list(set(family_email)))
-        result_string = (
-            result_string
-            + "\n\n\nHusk nu at bruge Bcc! ... IKKE TO: og heller IKKE CC: felterne\n\n"
-        )
-
-        return HttpResponse(result_string, content_type="text/plain")
-
-    export_emaillist.short_description = "Exporter e-mail liste"
-
-    def export_csv(self, request, queryset):
-        result_string = '"Navn";"Alder";"Opskrevet";"Tlf (barn)";"Email (barn)";"Tlf (forælder)";"Email (familie)";"Postnummer"\n'
-        for person in queryset:
-            parent = person.family.get_first_parent()
-            if parent:
-                parent_phone = parent.phone
-            else:
-                parent_phone = ""
-
-            if not person.family.dont_send_mails:
-                person_email = person.email
-                family_email = person.family.email
-            else:
-                person_email = ""
-                family_email = ""
-
-            result_string = (
-                result_string
-                + person.name
-                + ";"
-                + str(person.age_years())
-                + ";"
-                + str(person.added)
-                + ";"
-                + person.phone
-                + ";"
-                + person_email
-                + ";"
-                + parent_phone
-                + ";"
-                + family_email
-                + ";"
-                + person.zipcode
-                + "\n"
-            )
-            response = HttpResponse(result_string, content_type="text/csv")
-            response["Content-Disposition"] = 'attachment; filename="personer.csv"'
-        return response
-
-    export_csv.short_description = "Exporter CSV"
-
-    # Only view persons related to users department (all family, via participant, waitinglist & invites)
-    def get_queryset(self, request):
-        qs = super(PersonAdmin, self).get_queryset(request)
-        if request.user.is_superuser or request.user.has_perm(
-            "members.view_all_persons"
-        ):
-            return qs
-        else:
-            departments = Department.objects.filter(
-                adminuserinformation__user=request.user
-            ).values("id")
-            return qs.filter(
-                Q(
-                    family__person__member__activityparticipant__activity__department__in=departments
-                )
-                | Q(family__person__waitinglist__department__in=departments)
-                | Q(
-                    family__person__activityinvite__activity__department__in=departments
-                )
-            ).distinct()
-
-
-admin.site.register(Person, PersonAdmin)
-
-admin.site.register(EmailTemplate)
-
-
-class PaymentAdmin(admin.ModelAdmin):
-    list_display = [
-        "pk",
-        "added",
-        "payment_type",
-        "amount_ore",
-        "family",
-        "confirmed_dtm",
-        "cancelled_dtm",
-        "rejected_dtm",
-        "activityparticipant",
-    ]
-    list_filter = ["payment_type", "activity"]
-    raw_id_fields = ("person", "activityparticipant", "family")
-    date_hierarchy = "added"
-    search_fields = ("family__email",)
-    select_related = "activityparticipant"
-
-
-admin.site.register(Payment, PaymentAdmin)
-# admin.site.register(QuickpayTransaction)
 
 
 class EquipmentLoanInline(admin.TabularInline):
