+++ members/admin/person_admin_filters.py
@@ -1,7 +1,5 @@
 from django.contrib import admin
-from members.models import Activity
-
-from members.models import AdminUserInformation
+from members.models import Activity, AdminUserInformation
 
 
 class PersonParticipantListFilter(admin.SimpleListFilter):
@@ -10,13 +8,9 @@ class PersonParticipantListFilter(admin.SimpleListFilter):
 
     def lookups(self, request, model_admin):
         activitys = [("none", "Deltager ikke"), ("any", "Alle deltagere samlet")]
-        for activity in (
-            Activity.objects.filter(
-                department__in=AdminUserInformation.get_departments_admin(request.user)
-            )
-            .order_by("start_date")
-            .order_by("zipcode")
-        ):
+        for activity in Activity.objects.filter(
+            department__in=AdminUserInformation.get_departments_admin(request.user)
+        ).order_by("department__name", "-start_date"):
             activitys.append((str(activity.pk), str(activity)))
 
         return activitys
@@ -38,13 +32,9 @@ class PersonInvitedListFilter(admin.SimpleListFilter):
 
     def lookups(self, request, model_admin):
         activitys = [("none", "Ikke inviteret til noget"), ("any", "Alle inviterede")]
-        for activity in (
-            Activity.objects.filter(
-                department__in=AdminUserInformation.get_departments_admin(request.user)
-            )
-            .order_by("start_date")
-            .order_by("zipcode")
-        ):
+        for activity in Activity.objects.filter(
+            department__in=AdminUserInformation.get_departments_admin(request.user)
+        ).order_by("department__name", "-start_date"):
             activitys.append((str(activity.pk), str(activity)))
 
         return activitys
@@ -69,7 +59,9 @@ class PersonWaitinglistListFilter(admin.SimpleListFilter):
             ("any", "Alle opskrevne samlet"),
             ("none", "Ikke skrevet på venteliste"),
         ]
-        for department in AdminUserInformation.get_departments_admin(request.user):
+        for department in AdminUserInformation.get_departments_admin(
+            request.user
+        ).order_by("name"):
             departments.append((str(department.pk), department.name))
 
         return departments
@@ -91,7 +83,9 @@ class VolunteerListFilter(admin.SimpleListFilter):
 
     def lookups(self, request, model_admin):
         departments = [("any", "Alle frivillige samlet"), ("none", "Ikke frivillig")]
-        for department in AdminUserInformation.get_departments_admin(request.user):
+        for department in AdminUserInformation.get_departments_admin(
+            request.user
+        ).order_by("name"):
             departments.append((str(department.pk), department.name))
 
         return departments
